requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
convertdate==2.4.0
cryptography==42.0.5
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docutils==0.21.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
exchangelib==5.2.1
fonttools==4.51.0
fpdf2==2.7.5
furl==2.1.3
graphviz==0.13.2
h11==0.14.0
hijri-converter==2.3.1
holidays==0.21.13
htmldocx==0.0.6
idna==3.7
importlib_metadata==7.1.0
isodate==0.6.1
java-access-bridge-wrapper==1.2.0
jsonpath-ng==1.6.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
korean-lunar-calendar==0.3.1
lxml==5.2.1
more-itertools==10.2.0
mss==6.1.0
netsuitesdk==1.24.0
notifiers==1.3.3
O365==2.0.26
oauthlib==3.2.2
openpyxl==3.1.2
orderedmultidict==1.0.1
outcome==1.3.0.post0
packaging==24.0
pdfminer.six==20221105
pendulum==2.1.2
pillow==10.3.0
platformdirs==4.2.1
ply==3.11
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
PyMeeus==0.5.12
pynput-robocorp-fork==5.0.0
pyotp==2.9.0
pypdf==3.17.4
pyperclip==1.8.2
PySocks==1.7.1
pyspnego==0.10.2
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
python-xlib==0.33
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-file==2.0.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
robocorp-storage==1.0.4
robotframework==6.1.1
robotframework-pythonlibcore==4.4.1
robotframework-requests==0.9.7
robotframework-seleniumlibrary==6.3.0
robotframework-seleniumtestability==2.1.0
rpaframework==28.5.1
rpaframework-core==11.3.2
rpaframework-pdf==7.3.2
rpds-py==0.18.0
selenium==4.15.2
simple-salesforce==1.12.6
six==1.16.0
smartsheet-python-sdk==3.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stringcase==1.2.0
tenacity==8.2.3
trio==0.25.0
trio-websocket==0.11.1
tweepy==3.10.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.2.1
validators==0.20.0
webdriver-manager==4.0.1
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==4.2.1
zipp==3.18.1
